pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "llm-wiki-vnext"
version = "0.1.0"
description = "Agent-native compiled knowledge vault for enterprise teams."
requires-python = ">=3.11"
dependencies = [
  "alembic>=1.15.0",
  "anthropic>=0.49.0",
  "psycopg[binary]>=3.2.0",
  "asyncpg>=0.30.0",
  "beautifulsoup4>=4.12.3",
  "cryptography>=44.0.0",
  "email-validator>=2.2.0",
  "fastapi>=0.115.12",
  "httpx>=0.28.1",
  "mcp>=1.7.0",
  "minio>=7.2.15",
  "numpy>=2.2.5",
  "openai>=1.76.0",
  "opentelemetry-api>=1.25.0",
  "opentelemetry-exporter-otlp-proto-grpc>=1.25.0",
  "opentelemetry-instrumentation-fastapi>=0.46b0",
  "opentelemetry-sdk>=1.25.0",
  "openpyxl>=3.1.5",
  "orjson>=3.10.16",
  "bcrypt>=4.0.0",
  "prometheus-client>=0.21.0",
  "pydantic-settings>=2.8.1",
  "pydantic>=2.11.3",
  "PyJWT[crypto]>=2.10.1",
  "pypdf>=5.4.0",
  "python-docx>=1.1.2",
  "python-multipart>=0.0.20",
  "redis>=5.2.1",
  "slowapi>=0.1.9",
  "structlog>=24.4.0",
  "uvicorn[standard]>=0.34.2"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.5",
  "pytest-asyncio>=0.26.0",
  "pytest-env>=1.1.5",
  "ruff>=0.11.6"
]

[tool.setuptools]
packages = ["llm_wiki_core"]
package-dir = {"" = "shared/python"}

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
testpaths = ["tests"]
pythonpath = [".", "shared/python"]
env = [
  "D:INTERNAL_SERVICE_TOKEN=0123456789abcdef0123456789abcdef",
  "D:ACTIVE_KEY_VERSION=test-v1",
  "D:R:KEYRING_JSON={\"test-v1\":\"MDEyMzQ1Njc4OWFiY2RlZjAxMjM0NTY3ODlhYmNkZWY\"}",
]
